+++ 2D/DDLA.cpp
@@ -6,7 +6,7 @@ using namespace std;
 
 class DDLA : public CoarseImage {
 public:
-    explicit DDLA(int n_) : CoarseImage(n_, n_, H['f'] ? pow(n_, .33) : 1), f(H['f']), n(n_), cursum(0), p(H['p']) {
+    explicit DDLA(int n_) : CoarseImage(n_, n_, H['f'] ? int(pow(n_, .33)) : 1), f(H['f']), n(n_), cursum(0), p(H['p']) {
         put({0, 0}, true);
         pq.push({{1, 0}, prng.exponential() / p});
         pq.push({{0, 1}, prng.exponential() / (1 - p)});
@@ -58,7 +58,7 @@ public:
     PointQueue pq;
 };
 
-int main(int argc, char ** argv) {
+int main(int argc, char **argv) {
     H.init("Directed DLA", argc, argv, "n=750,p=.5,f");
     DDLA img(H['n']);
     img.show();
